# perf: improve CPF algorithm performance