The second CPF table can be completely elided, not requiring the algorithm
to iterate over it at all.

Ignoring the last digit on the second CPF table we have:

```
S1  = cpf[0] * 10 + cpf[1] *  9 ... cpf[9] * 2
S2' = cpf[0] * 11 + cpf[1] * 10 ... cpf[9] * 3
```

Each digit is multiplied by different weights in S1 and S2,
however the difference for each digit's weights is always +1.
Thus, the difference between S2' and S1 is simply the sum of the digits:

```
S2' - S1 = cpf[0] + cpf[1] + ... cpf[9]
S2' = S1 + cpf[0] + cpf[1] + ... cpf[9]
```

And then, to get the value of S2 we can just add 2 * cpf[10] to S2',
which is the last element of the second CPF table multiplied by the first
verifying digit of the CPF.

```
S2 = S2' + 2 * cpf[10]
```

```
goos: linux
goarch: amd64
pkg: github.com/phenpessoa/br
cpu: Intel(R) Core(TM) Ultra 7 155H
                        │   old.txt    │              new2.txt               │
                        │    sec/op    │   sec/op     vs base                │
GenerateCPF-21             44.02n ± 1%   38.90n ± 5%  -11.62% (p=0.000 n=10)
CPF_IsValid14-21          16.030n ± 1%   9.557n ± 1%  -40.38% (p=0.000 n=10)
CPF_IsValid11-21          12.610n ± 3%   7.927n ± 4%  -37.13% (p=0.000 n=10)
CPF_IsValid14Invalid-21   15.685n ± 2%   9.450n ± 1%  -39.75% (p=0.000 n=10)
CPF_IsValid11Invalid-21   12.530n ± 2%   7.915n ± 1%  -36.84% (p=0.000 n=10)
CPF_String14-21            16.68n ± 1%   10.29n ± 2%  -38.26% (p=0.000 n=10)
CPF_String11-21            27.50n ± 1%   23.41n ± 2%  -14.91% (p=0.000 n=10)
geomean                    18.71n        12.69n       -32.15%
```